// frontend/src/app/shared/enum/history-actions.enum.ts
export enum HistoryActions {
  REQUEST_CREATED = 'REQUEST_CREATED',
  QUOTE_CREATED = 'QUOTE_CREATED',
  QUOTE_APPROVED = 'QUOTE_APPROVED',
  QUOTE_REJECTED = 'QUOTE_REJECTED',
  REQUEST_RESCUED = 'REQUEST_RESCUED',
  MAINTENANCE_REDIRECTED = 'MAINTENANCE_REDIRECTED',
  MAINTENANCE_COMPLETED = 'MAINTENANCE_COMPLETED',
  PAYMENT_CONFIRMED = 'PAYMENT_CONFIRMED',
  REQUEST_FINALIZED = 'REQUEST_FINALIZED',
}